Fix network detection for Windows OS (#94)
* Fix Warnings

Fix some of the compile warnings on invalid castings

* Improve parallels detection

Parallels Desktop (MAC) is now properly detected and identified (not as Other)

* Fix network detection for Windows OS

- Proper usage of GetAdaptersInfo

+++ src/library/base/file_utils.cpp
@@ -37,7 +37,7 @@ string get_file_contents(const char *filename, size_t max_size) {
 	ifstream in(filename, std::ios::binary);
 	if (in) {
 		string contents;
-		size_t index = in.seekg(0, ios::end).tellg();
+		size_t index = (size_t)in.seekg(0, ios::end).tellg();
 		size_t limited_size = min(index, max_size);
 		contents.resize(limited_size);
 		in.seekg(0, ios::beg);
